migrated NotchCrack to Fracture, NotchCrack example to Fracture2DMode1
